fix(sdk): Support non-TLS hosts other than localhost
